Fix a bug in evaluation
The original code worked in a different way from its apparent intention, and thus discarded 840 perfectly correct sentences from evaluation. Effectively all sentences with no particles following the target word were discarded, which not only reduced the dataset drastically, but biased it as well.  Specifically, `r[-len(right):]` would be equivalent to the entire string `r`  if `len(right)` was zero, because `r[-0:]` equals `r[0:]` equals `r`. Consequently `right == r[-len(right):]` would evaluate as false for all cases where `len(right) == 0`. That's clearly not the original intention. Instead the new code replaces the condition with an assert that is in line with the apparent intention of the original code (the assert's condition is of course always true). Note: this change will have an effect on the evaluation results.